feat : 에러처리 구현 및 타입별로 파일 정리
