+++ Calculator/Calculator/Model/ErrorHandling.swift
@@ -1,0 +1,13 @@
+
+import Foundation
+
+enum QueueError: LocalizedError {
+    case EmptyInLinkedList
+    
+    var errorDescription: String? {
+        switch self {
+        case .EmptyInLinkedList: return "There is no Node"
+        }
+    }
+    
+}
