+++ Calculator/Calculator/Protocol/CalculateItem.swift
@@ -2,6 +2,6 @@
 import Foundation
 
 protocol CalculateItem {
-   
+
 }
